refactor(gh): CI用スクリプトをpackageとして整理
