+++ packages-private/README.md
@@ -1,0 +1,3 @@
+# packages-private
+
+`packages-private` は、CIなどで使用するためのユーティリティなどを格納する場所です。この配下にあるものは**プロダクションの `/packages`**の依存となるものではありません。
