+++ packages-private/changelog-checker/test/checker.test.ts
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import {expect, suite, test} from "vitest";
-import {Release, ReleaseCategory} from "../src/parser";
-import {checkNewRelease, checkNewTopic} from "../src/checker";
+import {expect, suite, test} from 'vitest';
+import {Release, ReleaseCategory} from '@/parser.js';
+import {checkNewRelease, checkNewTopic} from '@/checker.js';
 
 suite('checkNewRelease', () => {
 	test('headに新しいリリースがある1', () => {
