+++ packages-private/changelog-checker/vite.config.ts
@@ -1,0 +1,12 @@
+import {defineConfig} from 'vite';
+
+
+const config = defineConfig({
+	resolve: {
+		alias: {
+			'@': './src',
+		},
+	},
+});
+
+export default config;
