+++ packages-private/memory-usage-backend/package.json
@@ -1,0 +1,11 @@
+{
+	"name": "memory-usage-backend",
+	"private": true,
+	"type": "module",
+	"scripts": {
+		"run": "tsx src/index.ts"
+	},
+	"devDependencies": {
+		"tsx": "4.22.4"
+	}
+}
