--- scripts/changelog-checker/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-coverage
-.idea
