--- scripts/changelog-checker/vite.config.ts
@@ -1,6 +1,0 @@
-import {defineConfig} from 'vite';
-
-
-const config = defineConfig({});
-
-export default config;
